// src/api-client/models/location-pickup-slot-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LocationPickupSlotResponseDto
 */
export interface LocationPickupSlotResponseDto {
    /**
     * Unique slot ID.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'id': string;
    /**
     * Branch offering the slot.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'locationId': string;
    /**
     * What the slot is called to both sides.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'label': string;
    /**
     * ISO weekday, 1 = Monday. Null means every day.
     * @type {number}
     * @memberof LocationPickupSlotResponseDto
     */
    'weekday'?: number | null;
    /**
     * Opening time, branch-local.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'startTime': string;
    /**
     * Closing time, branch-local.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'endTime': string;
    /**
     * The window as one label, for display.
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'windowLabel': string;
    /**
     * Whether the slot is currently offered.
     * @type {boolean}
     * @memberof LocationPickupSlotResponseDto
     */
    'isActive': boolean;
    /**
     * When the slot was created (ISO 8601).
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'createdAt': string;
    /**
     * When the slot was last modified (ISO 8601).
     * @type {string}
     * @memberof LocationPickupSlotResponseDto
     */
    'updatedAt': string;
}